Taak 16.

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
  */
 int main(void)
 {
-	unsigned long long i, j, n;
+	unsigned long i, j, n;
 	int k;
 
 	i = 1;
